backend/users/pyproject.toml
[project]
name = "users"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiobotocore>=2.21.1",
    "aiokafka>=0.12.0",
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "black>=25.1.0",
    "cryptography>=44.0.2",
    "fastapi>=0.115.11",
    "loguru>=0.7.3",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.10.6",
    "pyjwt>=2.10.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.39",
    "strawberry-graphql>=0.262.5",
    "uvicorn>=0.34.0",
]
